chore: move rlp and trie folder to src
