--- src/ethrex/mod.rs
@@ -1,2 +1,0 @@
-pub mod rlp;
-pub mod trie;
