+++ src/trie/branch.rs
