+++ src/trie/error.rs
